Fix position rouding issue with go-to widget

+++ src/ui/go_to_widget.h
@@ -40,7 +40,7 @@ class GoToWidget : public QWidget
     void set_defaults(float default_x, float default_y);
 
   Q_SIGNALS:
-    void go_to_requested(int x, int y);
+    void go_to_requested(float x, float y);
 
   public Q_SLOTS:
